Refactor : normal bar 위치 계산 부분 공통 컴포넌트로 변경
