+++ src/jetty-chart/src/common/bar-common/utils/calculate-bar-positions.js
@@ -1,0 +1,195 @@
+export const calculateWarpperTransform = ({ horizontal, reverse, innerMargin, padding }) => {
+  return horizontal
+    ? `translate(${reverse ? innerMargin.top : innerMargin.bottom},${padding})`
+    : `translate(${padding}, ${reverse ? innerMargin.bottom : innerMargin.top})`;
+};
+
+export const calculateBarWrapperTransform = ({
+  useAnimation,
+  useTranslate,
+  renderType,
+  horizontal,
+  zeroHeight,
+  center,
+  halfBarRealWidth,
+  drawHeight,
+  barHeight
+}) => {
+  return useAnimation && useTranslate
+    ? "translate(0,0)"
+    : useAnimation && renderType.includes("grow")
+    ? horizontal
+      ? `translate(${zeroHeight},${center - halfBarRealWidth})`
+      : `translate(${center - halfBarRealWidth})`
+    : horizontal
+    ? `translate(${zeroHeight},${center - halfBarRealWidth})`
+    : `translate(${center - halfBarRealWidth},${drawHeight - barHeight - zeroHeight})`;
+};
+
+export const calculateBarWrapperFrom = ({
+  horizontal,
+  zeroHeight,
+  translate,
+  borderRadius,
+  center,
+  rectHeight,
+  rectWidth,
+  drawHeight,
+  barHeight
+}) => {
+  return horizontal
+    ? `${zeroHeight - translate.zeroHeight - borderRadius}px,${center - translate.center - (rectHeight - translate.height) / 2}px`
+    : `${center - translate.center - (rectWidth - translate.width) / 2}px,${drawHeight - barHeight - zeroHeight}px`;
+};
+
+export const calculateBarWrapperTo = ({ horizontal, zeroHeight, translate, center, drawHeight, barHeight, halfBarRealWidth }) => {
+  return horizontal
+    ? `${zeroHeight - translate.zeroHeight}px,${center - halfBarRealWidth}px`
+    : `${center - halfBarRealWidth}px,${drawHeight - barHeight - zeroHeight}px`;
+};
+
+export const calculateBarTransform = ({
+  useAnimation,
+  renderType,
+  useTranslate,
+  horizontal,
+  checkPositive,
+  barOnlyUpperRadius,
+  borderRadius,
+  barHeight
+}) => {
+  return (useAnimation && renderType.includes("grow")) || (useAnimation && useTranslate)
+    ? ""
+    : horizontal
+    ? `translate(${checkPositive ? (barOnlyUpperRadius ? -borderRadius : 0) : -barHeight})`
+    : `translate(0,${checkPositive ? 0 : barHeight - (barOnlyUpperRadius ? borderRadius : 0)})`;
+};
+
+export const calculateBarFrom = ({
+  useTranslate,
+  horizontal,
+  checkPositive,
+  borderRadius,
+  rectWidth,
+  translate,
+  barHeight,
+  barOnlyUpperRadius,
+  drawHeight,
+  zeroHeight
+}) => {
+  return useTranslate
+    ? horizontal
+      ? `${checkPositive ? 0 : borderRadius + borderRadius - rectWidth + translate.width}px`
+      : `0px,${(checkPositive ? translate.height : barHeight + (barOnlyUpperRadius ? -borderRadius : 0)) + translate.zeroHeight}px`
+    : horizontal
+    ? `${barOnlyUpperRadius ? (checkPositive ? -borderRadius : borderRadius) : 0}px,0px`
+    : `0px,${drawHeight - zeroHeight + (barOnlyUpperRadius ? (checkPositive ? borderRadius : -borderRadius) : 0)}px`;
+};
+
+export const calculateBarTo = ({
+  useTranslate,
+  horizontal,
+  checkPositive,
+  borderRadius,
+  rectWidth,
+  translate,
+  barHeight,
+  barOnlyUpperRadius,
+  drawHeight,
+  zeroHeight
+}) => {
+  return useTranslate
+    ? horizontal
+      ? `${(checkPositive ? -borderRadius : -rectWidth + borderRadius) + translate.zeroHeight}px`
+      : `0px,${checkPositive ? 0 : barHeight - borderRadius}px`
+    : horizontal
+    ? `${checkPositive ? (barOnlyUpperRadius ? -borderRadius : 0) : -barHeight}px,0px`
+    : `0px,${drawHeight - zeroHeight - (checkPositive ? barHeight : barOnlyUpperRadius ? borderRadius : 0)}px`;
+};
+
+export const calculateLabelTransform = ({
+  useAnimation,
+  useTranslate,
+  horizontal,
+  horizontalLabelLocation,
+  halfBarRealWidth,
+  verticalLabelLocation,
+  renderType,
+  drawHeight,
+  barHeight,
+  zeroHeight
+}) => {
+  return useAnimation && useTranslate
+    ? ``
+    : horizontal
+    ? `translate(${horizontalLabelLocation},${halfBarRealWidth})`
+    : `translate(${halfBarRealWidth},${
+        verticalLabelLocation + (useAnimation && renderType.includes("grow") ? drawHeight - barHeight - zeroHeight : 0)
+      })`;
+};
+
+/* eslint-disable complexity */
+export const calculateLabelFrom = ({
+  useTranslate,
+  horizontal,
+  labelPosition,
+  checkPositive,
+  rectWidth,
+  translate,
+  barBorderRadius,
+  labelMargin,
+  rectHeight,
+  borderRadius,
+  barHeight
+}) => {
+  return useTranslate
+    ? horizontal
+      ? labelPosition === "over"
+        ? `${(checkPositive ? rectWidth - translate.width : barBorderRadius) + labelMargin}px,${(rectHeight - translate.height) / 2}px`
+        : labelPosition === "under"
+        ? `${(checkPositive ? 0 : -rectWidth + translate.width + borderRadius) + borderRadius - labelMargin}px,${
+            (rectHeight - translate.height) / 2
+          }px`
+        : `${(checkPositive ? (barHeight - translate.width) / 2 : -(barHeight - translate.width) / 2) + borderRadius}px,${
+            (rectHeight - translate.height) / 2
+          }px`
+      : labelPosition === "over"
+      ? `${(rectWidth - translate.width) / 2}px,${(checkPositive ? translate.height : barHeight) - labelMargin + translate.zeroHeight}px`
+      : labelPosition === "under"
+      ? `${(rectWidth - translate.width) / 2}px,${
+          (checkPositive ? barHeight : barHeight + barHeight - translate.height) + labelMargin + translate.zeroHeight
+        }px`
+      : `${(rectWidth - translate.width) / 2}px,${
+          checkPositive
+            ? translate.height + (barHeight - translate.height) / 2 + translate.zeroHeight
+            : barHeight + (barHeight - translate.height) / 2 + translate.zeroHeight
+        }px`
+    : horizontal
+    ? labelPosition === "over"
+      ? `${checkPositive ? -barHeight : 0}px,0px`
+      : labelPosition === "under"
+      ? `${checkPositive ? 0 : barHeight}px,0px`
+      : `${checkPositive ? -barHeight / 2 : barHeight / 2}px,0px`
+    : labelPosition === "over"
+    ? `0px,${checkPositive ? barHeight : 0}px`
+    : labelPosition === "under"
+    ? `0px,${checkPositive ? 0 : -barHeight}px`
+    : `0px,${checkPositive ? barHeight / 2 : -barHeight / 2}px`;
+};
+/* eslint-enable complexity */
+
+export const calculateLabelTo = ({ useTranslate, horizontal, labelPosition, checkPositive, barHeight, labelMargin, translate, halfBarRealWidth }) => {
+  return useTranslate
+    ? horizontal
+      ? labelPosition === "over"
+        ? `${(checkPositive ? barHeight : 0) + labelMargin + translate.zeroHeight}px,${halfBarRealWidth}px`
+        : labelPosition === "under"
+        ? `${(checkPositive ? 0 : -barHeight) + translate.zeroHeight - labelMargin}px,${halfBarRealWidth}px`
+        : `${(checkPositive ? barHeight / 2 : -barHeight / 2) + translate.zeroHeight}px,${halfBarRealWidth}px`
+      : labelPosition === "over"
+      ? `${halfBarRealWidth}px,${(checkPositive ? 0 : barHeight) - labelMargin}px`
+      : labelPosition === "under"
+      ? `${halfBarRealWidth}px,${(checkPositive ? barHeight : barHeight + barHeight) + labelMargin}px`
+      : `${halfBarRealWidth}px,${checkPositive ? barHeight / 2 : barHeight + barHeight / 2}px`
+    : `0px,0px`;
+};
