--- src/jetty-chart/src/common/bar-common/utils/calculate-measure.js
@@ -1,47 +1,0 @@
-export const calculateBase = ({ horizontal, height, margin, width, scopeResult, autoScope, innerMargin, padding, length, barGap }) => {
-  const totalWidth = horizontal ? height - margin.bottom - margin.top : width - margin.left - margin.right;
-  const totalHeight = horizontal ? width - margin.left - margin.right : height - margin.bottom - margin.top;
-  const totalScope = scopeResult.maxScope - scopeResult.minScope;
-
-  if (!autoScope && scopeResult.display) {
-    innerMargin.top = scopeResult.topMarginRatio * totalHeight;
-    innerMargin.bottom = scopeResult.bottomMarginRatio * totalHeight;
-  }
-
-  const drawWidth = totalWidth - padding - padding;
-  const drawHeight = totalHeight - innerMargin.top - innerMargin.bottom;
-  const lineHeight = drawHeight / (scopeResult.scope.length - 1);
-
-  const barWidth = drawWidth / length;
-  const halfBarWidth = barWidth / 2;
-  const halfBarRealWidth = halfBarWidth - barGap * halfBarWidth;
-
-  const zeroHeight =
-    scopeResult.scope.reduce((acc, cur) => {
-      if (cur !== 0) {
-        acc += 1;
-      }
-
-      if (cur === 0) {
-        acc = 0;
-      }
-
-      return acc;
-    }, 0) * lineHeight;
-
-  return {
-    totalWidth,
-    totalHeight,
-    totalScope,
-    innerMargin,
-    drawWidth,
-    drawHeight,
-    lineHeight,
-    barWidth,
-    halfBarWidth,
-    halfBarRealWidth,
-    zeroHeight
-  };
-};
-
-export const calculateBarBase = () => {};
